Service files
* Add systemd scripts for Aphlict and Phrabricator daemons
* Add Aphlict and phab-phd systemd scripts to service script
* Remove references to deprecated variables
* Update README to reflect systemd script additions
This commit adds two systemd scripts for managing Aphlict and Phabricator's
daemons.

Their ExecStart locations are based on an where the appropriate files would
be in a Phabricator instance located in /usr/local.

* Add Aphlict and phab-phd systemd scripts to service script

The "service" script previously controlled Aphlict and the Phabricator
daemons by launching them manually - as you would from the terminal.

This commit replaces those lines with references to the systemd scripts
added in the previous commit. These systemd scripts should be placed in
/etc/systemd/system in order to be used properly.

* Remove references to deprecated variables

This commit removes references to "PHAB_PHD_USER" and "WEB_USER" from the
service script. These are no longer needed since the appropriate variables
have been moved to the systemd scripts for phab-phd and aphlict, respectively.

* Update README to reflect systemd script additions

This commit removes references to deprecated values in the service script
and adds a section referring to the new systemd scripts for Aphlict and phab-phd,
including how to manage them.